Full 0.1 version

+++ src/contracts/pointer_namespace.vala
@@ -9,10 +9,20 @@ namespace GData
 	 */
 	public class PointerNamespace : Object, GLib.ListModel
 	{
+		private bool dirty = false;
+		private bool cleaning = false;
+
 		private GLib.Array<WeakRefWrapper> _pointers = new GLib.Array<WeakRefWrapper>();
 
 		private static PointerNamespace? _instance = null;
 
+		/**
+		 * Returns singleton instance of binding namespace
+		 * 
+		 * @since 0.1
+		 * 
+		 * @return Instance of binding namespace
+		 */
 		public static PointerNamespace get_instance()
 		{
 			if (_instance == null)
@@ -23,11 +33,21 @@ namespace GData
 		// Safer to just clean everything as it is small amount of data
 		private void clean_null()
 		{
+			if (cleaning == true) {
+				dirty = true;
+				return;
+			}
+			cleaning = true;
 			for (int i=(int)_pointers.length-1; i>=0; i--) {
 				if (((WeakRefWrapper) _pointers.data[(uint)i]).target == null) {
 					_pointers.remove_index ((uint)i);
 					items_changed ((uint)i, 1, 0);
 				}
+			}
+			cleaning = false;
+			if (dirty == true) {
+				dirty = false;
+				clean_null();
 			}
 		}
 
@@ -37,6 +57,14 @@ namespace GData
 			items_changed (_pointers.length, 0, 1);
 		}
 
+		/**
+		 * Resolves pointer by specified id
+		 * 
+		 * @since 0.1
+		 * 
+		 * @param id Pointer id
+		 * @return Pointer reference if found, null if not
+		 */
 		public BindingPointer? get_by_id (int id)
 		{
 			for (uint i=0; i<_pointers.length; i++) {
@@ -44,20 +72,42 @@ namespace GData
 					if (as_pointer(get_item(i)).id == id)
 						return (as_pointer(get_item(i)));
 			}
-			stdout.printf("id(%i) not found\n", id);
+			GLib.warning ("id(%i) not found\n", id);
 			return (null);
 		}
 
+		/**
+		 * Returns pointer at position
+		 * 
+		 * @since 0.1
+		 * 
+		 * @param position Item position
+		 * @return Binding at specified position
+		 */
 		public Object? get_item (uint position)
 		{
 			return (((WeakRefWrapper) _pointers.data[position]).target);
 		}
 
+		/**
+		 * Returns item type
+		 * 
+		 * @since 0.1
+		 * 
+		 * @return Stored item type
+		 */
 		public Type get_item_type ()
 		{
 			return (typeof(BindingPointer));
 		}
-		
+
+		/**
+		 * Returns number of stored pointers
+		 * 
+		 * @since 0.1
+		 * 
+		 * @return Number of stored bindings
+		 */
 		public uint get_n_items ()
 		{
 			return (_pointers.length);
@@ -68,3 +118,4 @@ namespace GData
 		}
 	}
 }
+
